initial implementation of navbar with search on client side

+++ server/erp ecommerce/Models/ProductDto.cs
@@ -9,7 +9,7 @@
         public int? SizeId { get; set; }
         public int? ColorId { get; set; }
         public decimal? Price { get; set; }
-        public decimal? Discount { get; set; }
+        public int? Discount { get; set; }
         public int? CategoryId { get; set; }
         public int? BrandId { get; set; }
         public string ProductType { get; set; }
